updating top five container tanstack query and spending over time too
