+++ src/js/features/state/topFive/containers/TopFiveContainer.jsx
@@ -3,12 +3,10 @@
  * Created by Kevin Li 5/15/18
  */
 
-import React, { useCallback, useMemo, useState, useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import PropTypes from "prop-types";
 
-import { getTrailingTwelveMonths, convertFYToDateRange } from 'helpers/fiscalYearHelper';
-import { awardTypeGroups } from 'dataMapping/search/awardType';
 import TopFive from "components/sharedComponents/TopFive";
 import useFetchSpendingBy from "./useFetchSpendingBy";
 
@@ -25,74 +23,9 @@ const TopFiveContainer = ({ category, type, agencyData }) => {
 
     const { code, _totalAmount: total } = overview;
 
-    const getDataParams = useCallback(() => {
-        let timePeriod = null;
-        if (fy === 'latest') {
-            const trailing = getTrailingTwelveMonths();
-            timePeriod = {
-                start_date: trailing[0],
-                end_date: trailing[1]
-            };
-        }
-        else if (fy !== 'all' && fy) {
-            const range = convertFYToDateRange(parseInt(fy, 10));
-            timePeriod = {
-                start_date: range[0],
-                end_date: range[1]
-            };
-        }
-
-        const filters = {
-            place_of_performance_scope: 'domestic',
-            place_of_performance_locations: [
-                {
-                    country: 'USA',
-                    state: code
-                }
-            ]
-        };
-
-        if (timePeriod) {
-            filters.time_period = [timePeriod];
-        }
-
-        // Tab selection
-        if (type !== 'all' && awardTypeGroups[type]) {
-            filters.award_type_codes = awardTypeGroups[type];
-        }
-
-        const params = {
-            filters,
-            category,
-            limit: 5,
-            page: 1
-        };
-
-        if (category === 'awards') {
-            filters.award_type_codes = ['A', 'B', 'C', 'D'];
-            params.fields = ['Award ID', 'Award Amount', 'generated_internal_id'];
-            params.order = 'desc';
-            params.sort = 'Award Amount';
-            params.spending_level = 'awards';
-        }
-
-        if (category === 'defc') {
-            params.spending_level = 'award_financial';
-            params.filters = {
-                def_codes: ["L", "M", "N", "O", "P", "U", "V", "Z", "1"],
-                ...filters
-            };
-        }
-
-        return params;
-    }, [category, code, fy, type]);
-
-    const dataParams = useMemo(() => getDataParams(), [getDataParams]);
-
-
     const {
-        parsedData, noResults, isSuccess, isLoading, error
-    } = useFetchSpendingBy(dataParams, category, code, fy);
+        parsedData, noResults, isSuccess, isLoading, error, dataParams
+    } = useFetchSpendingBy(category, code, fy, type);
 
     useEffect(() => {
         if (isSuccess && (noResults || parsedData?.length > 0)) {
