+++ src/js/features/state/topFive/containers/useFetchSpendingBy.jsx
@@ -3,14 +3,80 @@
  * Created by Andrea Blackwell 03/19/26
  */
 
-import { useState, useEffect } from "react";
-import { useMutation } from '@tanstack/react-query';
+import { useState, useEffect, useCallback, useMemo } from "react";
+import { useQuery } from '@tanstack/react-query';
 import { performSpendingByAwardSearch, performSpendingByCategorySearch } from "helpers/searchHelper";
 import BaseStateCategoryResult from "models/v2/state/BaseStateCategoryResult";
+import { convertFYToDateRange, getTrailingTwelveMonths } from "../../../../helpers/fiscalYearHelper";
+import { awardTypeGroups } from "../../../../dataMapping/search/awardType";
 
-export const useFetchSpendingBy = (apiParams, category, code, fy) => {
+export const useFetchSpendingBy = (category, code, fy, type) => {
     const [parsedData, setParsedData] = useState(null);
     const [noResults, setNoResults] = useState(false);
+
+    const getDataParams = useCallback(() => {
+        let timePeriod = null;
+        if (fy === 'latest') {
+            const trailing = getTrailingTwelveMonths();
+            timePeriod = {
+                start_date: trailing[0],
+                end_date: trailing[1]
+            };
+        }
+        else if (fy !== 'all' && fy) {
+            const range = convertFYToDateRange(parseInt(fy, 10));
+            timePeriod = {
+                start_date: range[0],
+                end_date: range[1]
+            };
+        }
+
+        const filters = {
+            place_of_performance_scope: 'domestic',
+            place_of_performance_locations: [
+                {
+                    country: 'USA',
+                    state: code
+                }
+            ]
+        };
+
+        if (timePeriod) {
+            filters.time_period = [timePeriod];
+        }
+
+        // Tab selection
+        if (type !== 'all' && awardTypeGroups[type]) {
+            filters.award_type_codes = awardTypeGroups[type];
+        }
+
+        const params = {
+            filters,
+            category,
+            limit: 5,
+            page: 1
+        };
+
+        if (category === 'awards') {
+            filters.award_type_codes = ['A', 'B', 'C', 'D'];
+            params.fields = ['Award ID', 'Award Amount', 'generated_internal_id'];
+            params.order = 'desc';
+            params.sort = 'Award Amount';
+            params.spending_level = 'awards';
+        }
+
+        if (category === 'defc') {
+            params.spending_level = 'award_financial';
+            params.filters = {
+                def_codes: ["L", "M", "N", "O", "P", "U", "V", "Z", "1"],
+                ...filters
+            };
+        }
+
+        return params;
+    }, [category, code, fy, type]);
+
+    const dataParams = useMemo(() => getDataParams(), [getDataParams]);
 
     const parseData = (res) => {
         if (!res) {
@@ -58,29 +124,26 @@ export const useFetchSpendingBy = (apiParams, category, code, fy) => {
     };
 
     const {
-        mutate, isSuccess, isPending, error
-    } = useMutation({
-        mutationKey: [`spendingBy${category}${code}${fy}`],
-        mutationFn: () => {
+        data, isSuccess, isLoading, error
+    } = useQuery({
+        queryKey: [`spendingBy${category}${type}${code}${fy}`],
+        queryFn: () => {
             if (category === 'awards') {
-                return performSpendingByAwardSearch(apiParams).promise;
+                return performSpendingByAwardSearch(dataParams).promise;
             }
-            return performSpendingByCategorySearch(apiParams).promise;
-        },
-        onSuccess: (data) => {
-            parseData(data?.data);
+            return performSpendingByCategorySearch(dataParams).promise;
         },
         staleTime: 60000
     });
 
     useEffect(() => {
-        if (code && category) {
-            mutate();
+        if (isSuccess && Object.keys(data?.data).length > 0) {
+            parseData(data?.data);
         }
-    }, [code, mutate, category, apiParams]);
+    }, [data, isSuccess]);
 
     return {
-        parsedData, noResults, isSuccess, isLoading: isPending, error
+        parsedData, noResults, isSuccess, isLoading, error, dataParams
     };
 };
